requirements.txt
python-telegram-bot
python-dotenv
yt-dlp
requests
gofile2
colorlog 